Further refactoring of Score
